server debug

+++ compat/rdma.c
@@ -193,7 +193,7 @@ int iperf_cq_event_handler(struct rdma_cb *cb)
 	struct ibv_recv_wr *bad_wr;
 	int ret;
 	
-	DPRINTF(("cq_handler sem_wait @ 0x%", PRIx64, "\n", ntohll(&cb->sem)));
+	DPRINTF(("cq_handler sem_wait @ 0x%", PRIx64, "\n", ntohll((uint64_t)&cb->sem)));
 
 	while ((ret = ibv_poll_cq(cb->cq, 1, &wc)) == 1) {
 		ret = 0;
