+++ src/Client.cpp
@@ -62,6 +62,8 @@
 #include "Locale.h"
 #include "rdma.h"
 
+#include <inttypes.h>
+
 /* -------------------------------------------------------------------
  * Store server hostname, optionally local hostname, and socket info.
  * ------------------------------------------------------------------- */
@@ -284,7 +286,7 @@ void Client::RunRDMA( void ) {
 	DPRINTF(("client ibv_post_send success\n"));
 
 	/* Wait for server to ACK read complete */
-	DPRINTF(("sem_wait @ 0x%", PRIx64, "\n", ntohll(&mCb->sem)));
+	DPRINTF(("sem_wait @ 0x%", PRIx64, "\n", ntohll((uint64_t)&mCb->sem)));
 	sem_wait(&mCb->sem);
 	if (mCb->state != RDMA_WRITE_ADV) {
 		fprintf(stderr, "wait for RDMA_WRITE_ADV state %d\n",
